remove: 검증 로직을 클래스로 분리하지 않고 서비스 로직에 통합
